feat: add db connection check on startup & generalize other startup logic
